# openapi.yaml
openapi: 3.0.0
servers:
  - url: http://localhost8080
info:
  description: Contract API For The Skill-Matrix Application. This API contains all the neccessary endpoints for CRUD operations based on project requirements. In a nutshell, the application for which this API was created, contains all the active employees currently working for SoftwareONE. The purpose of this application is to aggregate all employees in one place so that any manager can access and contact any employee based on skills and other criterias.
  version: "1.0.0"
  title: Skill-Matrix API 
  contact:
    email: zourab.lavrentiadis@softwareone.com
  license:
    name: SoftwareONE
    url: 'http://www.softwareone.com'
...
paths:
...
  /language:
    get:
      tags:
        - language
      summary: fetch all languages from the database
      operationId: getLanguages
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Language'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server

    post:
      tags:
        - language
      summary: add a language to the database
      operationId: addLanguage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Language'
      responses:
        201:
           description: created
           content:
            application/json:
              schema:
                $ref: '#/components/schemas/Language'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server
  /language/{languageId}:
    get:
      tags:
        - language
      summary: find language by id
      operationId: getLanguageById
      parameters:
        - name: languageId
          in: path
          description: id of language to return
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Language'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server
 
    put:
      tags:
        - language
      summary: update a language by id
      operationId: updateLanguage
      parameters:
      - in: path
        name: languageId
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/Language'
      responses:
        201:
          description: created
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server
          
    delete:
      tags:
        - language
      summary: delete a language by id
      operationId: deleteLanguage
      parameters:
      - in: path
        name: languageId
        required: true
        schema:
          type: integer
      responses:
        201:
          description: deleted
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server  
        

 #language-level
  /language-level:
    get:
      tags:
        - language-level
      summary: fetch all language levels from the database
      operationId: getLanguageLevel
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LanguageLevel'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server  

    post:
      tags:
        - language-level
      summary: add a language level to the database
      operationId: addLanguageLevel
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LanguageLevel'
      responses:
        201:
           description: created
           content:
            application/json:
              schema:
                $ref: '#/components/schemas/LanguageLevel'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server  
  /language-level/{language-levelId}:
    put:
      tags:
        - language-level
      summary: update a language level
      operationId: updateLanguageLevel
      parameters:
      - in: path
        name: language-levelId
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/LanguageLevel'
      responses:
        201:
          description: created
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server  
          
    delete:
      tags:
        - language-level
      summary: delete a language-level
      operationId: deleteLanguageLevel
      parameters:
      - in: path
        name: language-levelId
        required: true
        schema:
          type: integer
      responses:
        201:
          description: deleted
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server    


 #job-title
...
components:
  schemas:
...
    Language:
      type: object
      properties:
        languageId:
          type: integer
          example: 1
        languageName:
          type: string
          enum:
            - english
            - german
            - greek
    LanguageLevel:
      type: object
      properties:
        LanguageLevelId:
          type: integer
          example: 1
        languageLevel:
          type: string
          enum:
            - a1
            - a2
            - b1
            - b2
            - c1
            - c2
...